jööti
XDDD

+++ app/src/main/java/com/example/fantasyquestwalker/Journey.java
@@ -2,16 +2,16 @@ package com.example.fantasyquestwalker;
 
 public class Journey {
     private String nimi;
-    private int matka;
+    private float matka;
 
-        public Journey(String nimi, int matka){
+        public Journey(String nimi, float matka){
             this.nimi = nimi;
             this.matka = matka;
         }
 
     public String getNimi() { return nimi; }
 
-    public int getMatka(){ return matka; }
+    public float getMatka(){ return matka; }
 
     @Override
     public String toString(){ return nimi; }
